refactor: Update celestial overlay markers
Replaced hardcoded SVG path generation with exported constants and
helper functions for `PLASTER_RING_CLIP_PATH` and `SoftBlurFilter`.

This update simplifies `CelestialMarkersOverlay` by removing local path
generation and utilizing shared SVG definition components, improving
maintainability.

The logic in `SunArcOverlay` was significantly updated to correctly
render the sun arc glow and core path using these new primitives,
especially adjusting path generation when the sun is visible during the
day.
